+++ src/entity/monster_nest.cpp
@@ -2,8 +2,7 @@
 
 MonsterNest::MonsterNest(int ix_, int iy_)
     : Structure(ix_, iy_, g_game.ingameMonsterNestSize, g_game.ingameMonsterNestSize),
-      MonsterSpawnNode(
-          g_game.ingameMonsterNestSize * 0.5f + ix_, g_game.ingameMonsterNestSize * 0.5f + iy_)
+      MonsterSpawnNode(g_game.ingameMonsterNestSize * 0.5f + ix_, g_game.ingameMonsterNestSize * 0.5f + iy_)
 {
     spawnX = nodeX;
     spawnY = nodeY;